requirements.txt
# Core dependencies for Treasury Risk Dashboard
aiohttp>=3.8.0
pandas>=1.5.0
numpy>=1.24.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# Database
asyncpg>=0.29.0
sqlalchemy>=2.0.0

# Data sources
yfinance>=0.2.0

# Machine Learning & Forecasting
scikit-learn>=1.3.0
statsmodels>=0.14.0
prophet>=1.1.0
torch>=2.0.0

# Time series analysis
timescaledb>=0.1.0

# API and networking
httpx>=0.25.0
requests>=2.31.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Logging and monitoring
structlog>=23.0.0